// js_ts/erp-crm/src/app/pages/sales-orders/interfaces/SalesOrders.interface.ts
export  interface ISales {
    key: React.Key;
    // number: number;
    date: string;
    employee: string;
    storage: string;
    client: string;
    summ: number;
    status: string;
  }
  
 export  interface IOrders {
    key: React.Key;
    // number: number;
    date: string;
    employee: string;
    storage: string;
    receipt: string;
    status: string;
  }